fix(driver): update b91, tl321x and tl323x drivers feat(driver): add tl521x platform

+++ tl_zigbee_sdk/proj/drivers/drv_spi.c
@@ -40,8 +40,8 @@ void drv_spi_master_init(u32 spiClock, drv_spi_mode_e mode)
 {
 #if defined(MCU_CORE_826x) || defined(MCU_CORE_8258) || defined(MCU_CORE_8278) || defined(MCU_CORE_B91)
     u8 divClock = ((2 * spiClock) >= SPI_CLOCK_SOURCE) ? 0 : (u8)(SPI_CLOCK_SOURCE / (2 * spiClock) - 1);
-#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X)
-    u8 divClock = SPI_CLOCK_SOURCE / spiClock;
+#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X) || defined(MCU_CORE_TL521X)
+    u16 divClock = SPI_CLOCK_SOURCE / spiClock;
 #endif
 
 #if defined(MCU_CORE_826x)
@@ -54,7 +54,7 @@ void drv_spi_master_init(u32 spiClock, drv_spi_mode_e mode)
      */
     spi_master_init(PSPI_MODULE, divClock, mode);
     spi_master_config(PSPI_MODULE, SPI_NORMAL);
-#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X)
+#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X) || defined(MCU_CORE_TL521X)
     spi_master_init(GSPI_MODULE, divClock, mode);
     spi_master_config(GSPI_MODULE, SPI_NORMAL);
 #endif
@@ -75,7 +75,7 @@ void drv_spi_slave_init(drv_spi_mode_e mode)
     spi_slave_init(0, mode);
 #elif defined(MCU_CORE_B91)
     spi_slave_init(PSPI_MODULE, mode);
-#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X)
+#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X) || defined(MCU_CORE_TL521X)
     spi_slave_init(GSPI_MODULE, mode);
 #endif
 }
@@ -116,7 +116,7 @@ void drv_spi_master_pin_select(pspi_clk_pin_def_e sclk_pin, pspi_csn_pin_def_e c
 
     pspi_set_pin(&pspi_pin_config);
 }
-#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X)
+#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X) || defined(MCU_CORE_TL521X)
 void drv_spi_master_pin_select(gpio_pin_e sclk_pin, gpio_pin_e cs_pin, gpio_pin_e mosi_pin, gpio_pin_e miso_pin)
 {
     gspi_pin_config_t gspi_pin_config;
@@ -166,7 +166,7 @@ void drv_spi_slave_pin_select(pspi_clk_pin_def_e sclk_pin, pspi_csn_pin_def_e cs
 
     pspi_set_pin(&pspi_pin_config);
 }
-#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X)
+#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X) || defined(MCU_CORE_TL521X)
 void drv_spi_slave_pin_select(gpio_pin_e sclk_pin, gpio_pin_e cs_pin, gpio_pin_e mosi_pin, gpio_pin_e miso_pin)
 {
     drv_spi_master_pin_select(sclk_pin, cs_pin, mosi_pin, miso_pin);
@@ -212,7 +212,7 @@ void drv_spi_write(u8 *cmd, int cmdLen, u8 *data, int dataLen, u32 csPin)
         spi_master_write(PSPI_MODULE, pBuf, cmdLen + dataLen);
         ev_buf_free(pBuf);
     }
-#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X)
+#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X) || defined(MCU_CORE_TL521X)
     u8 *pBuf = (u8 *)ev_buf_allocate(cmdLen + dataLen);
     if (pBuf) {
         u8 *pData = pBuf;
@@ -254,7 +254,7 @@ void drv_spi_read(u8 *cmd, int cmdLen, u8 *data, int dataLen, u32 csPin)
         drv_cs_pin = csPin;
     }
     spi_master_write_read(PSPI_MODULE, cmd, cmdLen, data, dataLen);
-#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X)
+#elif defined(MCU_CORE_TL321X) || defined(MCU_CORE_TL323X) || defined(MCU_CORE_TL521X)
     spi_master_write_read(GSPI_MODULE, cmd, cmdLen, data, dataLen);
 #endif
 }
